Progress update. Reworking test documentation. See #29.
