+++ src/journal/entry.rs
@@ -1,34 +1,5 @@
-
 use std::time::SystemTime;
 
-/// ##### Enumeration of journal entry severities as bytes flags in order of severity.
-#[allow(non_snake_case)]
-pub mod KJournalEntrySeverity {
-
-    /// Lowest severity. Debug message. Listeners shouldn't listener to those by default.
-    pub const DEBUG: u8 = 1;
-
-    /// Other message with low severity impact.
-    pub const OTHER: u8 = 2;
-
-    /// Information about event as they occur.
-    pub const INFORMATION: u8 = 4;
-
-    /// An occurred event that can be potentially severe.
-    pub const WARNING: u8 = 8;
-
-    /// A severe error that occurred but didn't cause the program to crash.
-    pub const ERROR: u8 = 16;
-
-    /// Highest severity. Causes program to crash.
-    pub const FATAL: u8 = 32;
-    
-    /// Quick shortcut to all severity flags without DEBUG.
-    pub const ALL_NO_DEBUG: u8 = 62;
-
-    /// Quick shortcut to all severity flags including DEBUG.
-    pub const ALL_WITH_DEBUG: u8 = 63;
-}
 
 /// ##### Journal entry. Contains the severity, the date and time of entry and a description.
 /// 
@@ -44,7 +15,7 @@ pub mod KJournalEntrySeverity {
 /// j.update(KJournalEntrySeverity::INFORMATION, String::from("This is now a recycle entry!"));
 /// ```
 pub struct KJournalEntry {
-    /// Severity of the entry according to [`KJournalEntrySeverity`].
+    /// Severity of the entry according to [`super::KJournalEntrySeverity`].
     severity : u8,
 
     /// Date and time entry occurred.
@@ -59,7 +30,7 @@ impl KJournalEntry {
     /// Create a new [`KJournalEntry`] from severity and description. Date and time will be added automatically.
     /// 
     /// # Argument(s)
-    /// * `severity` - [`KJournalEntrySeverity`] of the entry as u8.
+    /// * `severity` - [`super::KJournalEntrySeverity`] of the entry as u8.
     /// * `description` - Description or metadata of entry.
     /// 
     /// # Return
@@ -71,7 +42,7 @@ impl KJournalEntry {
     /// Update Journal entry with a new severity and description. Date and time will be modified automatically.
     /// 
     /// # Argument(s)
-    /// * `severity` - [`KJournalEntrySeverity`] of the entry as u8.
+    /// * `severity` - [`super::KJournalEntrySeverity`] of the entry as u8.
     /// * `description` - Description or metadata of entry.
     pub fn update(&mut self, severity : u8, description : String){
         self.date_time = SystemTime::now();
@@ -79,10 +50,10 @@ impl KJournalEntry {
         self.description = description;
     }
 
-    /// Get entry [`KJournalEntrySeverity`].
+    /// Get entry [`super::KJournalEntrySeverity`].
     /// 
     /// # Return
-    /// Entry [`KJournalEntrySeverity`].
+    /// Entry [`super::KJournalEntrySeverity`].
     pub fn get_severity(&self) -> u8{
         self.severity
     }
