+++ src/journal/listener.rs
@@ -2,16 +2,24 @@ use super::KJournalEntry;
 
 /// Implementing this trait permit to listen to journal
 /// 
+/// Listeners are notified of new journal entry according to severity they listen
+/// to according to set_severity().
 pub trait KJournalListener {
     /// Notification of new entry with an unmutable reference to it.
     fn notify(&self, new_entry : &KJournalEntry);
+
+    /// Set the severity the listener will listen to.
+    fn set_severity(&mut self, severity:u8);
+
+    /// Get the severity the listener is listening to.
+    fn get_severity(&self) -> u8;
 }
 
 
 /// List of listeners listening to the journal.
 pub struct KJournalListenerList<'a> {
 
-    listeners : Vec<KJournalListenerListEntry<'a>>
+    listeners : Vec<&'a dyn KJournalListener>
 
 }
 
@@ -20,17 +28,31 @@ impl<'a> KJournalListenerList<'a> {
     /// Create a new instance of KJournalListenerList
     pub fn new() -> KJournalListenerList<'a>{
         // Create listeners vector.
-        let mut listeners : Vec<KJournalListenerListEntry> = Vec::new();
+        let mut listeners : Vec<&'a dyn KJournalListener> = Vec::new();
 
         // Create List instance
         KJournalListenerList { listeners }
     }
+
+    pub fn notify(&self, new_entry : &super::KJournalEntry) {
+        todo!()
+    }
+
+    pub fn add_listener(&mut self, listener : &dyn KJournalListener) {
+        todo!()
+    }
+
+    pub fn remove_listener(&mut self, listener : &dyn KJournalListener) {
+        todo!()
+    }
+
+    pub fn count(&self)->usize {
+        todo!()
+    }
+
+    pub fn clear(&self) {
+        todo!()
+    }
 }
 
-/// Entry of a listener in the list.
-pub struct KJournalListenerListEntry<'a> {
-    /// Listener to notify.
-    listener : &'a dyn KJournalListener,
-    /// Severity(ies) this listener listen to.
-    severity : u8
-}
+
