+++ tests/integration_test.rs
@@ -1,2 +1,13 @@
+// Test header template
+
+/// # Test
+/// {FUNCTION NAME}
+/// 
+/// # Description
+/// {SHORT DESCRIPTION OF TEST}
+/// 
+/// # Verification(s)
+/// V{#} | {ENUMERATION OF ELEMENT(S) VERIFIED}
+
 // Contains Kleio tests modules
 pub mod kleio;
