+++ tests/kleio/journal/mod.rs
@@ -2,6 +2,14 @@
 #[cfg(test)]
 pub mod entry;
 
+// Contains tests for KJournalBuffer
+#[cfg(test)]
+pub mod buffer;
+
+// Contains tests for KJournalListenerList and KJournalListenerPrint
+#[cfg(test)]
+pub mod listener;
+
 // Contains tests for KJournal
 #[cfg(test)]
 pub mod journal;
